+++ crates/forge/tests/e2e/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod common;
 mod file_operations;
 mod fork;
 mod running;
+mod workspaces;