+++ CHANGELOG.md
@@ -17,6 +17,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Forge
 
 #### Added
+
+- Support for scarb workspaces.
+
+### Changed
+
+- Tests are collected only from `tests` folder and a package tree
+
+## [0.6.0] - 2023-09-13
+
+### Forge
+
+#### Added
+
 - `deploy_at` cheatcode
 - printing failures summary at the end of an execution
 - filtering tests now uses an absolute module tree path — it is possible to filter tests by module names, etc.